pyproject.toml
[tool.poetry]
name = "pubgpt"
version = "0.1.0"
description = "Extract relevant informations in PubMed papers in a simple way."
authors = ["Danilo Palumbo"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.23.1"
bio = "^1.5.9"
pypdf2 = "^3.0.1"
langchain = "^0.0.198"
python-dotenv = "^1.0.0"
cohere = "^4.9.0"
faiss-cpu = "^1.7.4"
openai = "^1.3.5"
tiktoken = "^0.5.1"



[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pylint = "^2.17.4"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
